+++ src/utils/base64.h
@@ -10,7 +10,7 @@ namespace EGL3::Utils {
 
 	std::string B64Encode(const std::string& Input);
 
-	std::string B64Decode(const uint8_t* Input, uint32_t InputSize);
+	std::string B64Decode(const uint8_t* Input, size_t InputSize);
 
-	std::string B64Encode(const uint8_t* Input, uint32_t InputSize);
+	std::string B64Encode(const uint8_t* Input, size_t InputSize);
 }
